==> 04_Generation_Social_Data.ipynb <==
# Generierung der sozialen Daten (Wochenende, Wochentag etc.)

**Libraries**<br>
pandas:     Datenverarbeitung<br>
os:         Betriebsystem-Funktionen für relative Pfadreferenzierung<br>

import pandas as pd
import os

**Auswahl des Datensatz (bereinigt oder unbereinigt)**

dirname = os.path.abspath('')

#path_merged_data =    os.path.join(dirname, '2019_Merged_Data_1h.csv')
path_merged_data =    os.path.join(dirname, '2019_Merged_Clean_Data_1h.csv')

merged_data =    pd.read_csv(path_merged_data)

# Function to map weekday (Monday as 1, Tuesday as 2, ..., Sunday as 7)
def map_weekday(timestamp):
    return timestamp.weekday() + 1

# Function to map weekend (Saturday (6) and Sunday (7) are considered weekends)
def map_weekend(day):
    return 1 if day in [6, 7] else 0

# Function to extract minute of the hour
def extract_minute(timestamp):
    return timestamp.minute

# Function to extract hour of the day
def extract_hour(timestamp):
    return timestamp.hour

# Function to extract month as an integer (January = 1, February = 2, ..., December = 12)
def extract_month(timestamp):
    return timestamp.month

# Assuming 'MESS_DATUM' column contains timestamp values
merged_data['MESS_DATUM'] = pd.to_datetime(merged_data['MESS_DATUM'])

# Create a new column 'Weekday' by mapping the weekday using the map_weekday function
merged_data['Weekday'] = merged_data['MESS_DATUM'].apply(lambda x: map_weekday(x))
# Create a new column 'Weekend' by mapping the weekend using the map_weekend function
merged_data['Weekend'] = merged_data['Weekday'].apply(lambda x: map_weekend(x))
# Create a new column 'Daytime' to store the extracted daytime information in integer format
merged_data['Month'] = merged_data['MESS_DATUM'].apply(lambda x: extract_month(x))
# Create a new column 'Daytime' to store the extracted hour of the day as integers
merged_data['Hour_of_Day'] = merged_data['MESS_DATUM'].apply(lambda x: extract_hour(x))
# Create a new column 'Minute_of_Hour' to store the extracted minute of the hour
#predictors_df['Minute_of_Hour'] = predictors_df['rec_time'].apply(lambda x: extract_minute(x))

print(merged_data)

merged_data.to_csv('2019_Trainingsdaten_1h.csv', index=False) #index=FALSE for not including row indices
